+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/home/minh/Desktop/study-management/src/index.tsx":"1","/home/minh/Desktop/study-management/src/reportWebVitals.ts":"2","/home/minh/Desktop/study-management/src/App.tsx":"3","/home/minh/Desktop/study-management/src/components/Users.tsx":"4","/home/minh/Desktop/study-management/src/layouts/index.tsx":"5","/home/minh/Desktop/study-management/src/components/shared/button/PrimaryBtn.tsx":"6","/home/minh/Desktop/study-management/src/graphql/autogenerate/hooks.tsx":"7","/home/minh/Desktop/study-management/src/layouts/nav.tsx":"8","/home/minh/Desktop/study-management/src/components/adminNav.tsx":"9"},{"size":845,"mtime":1612863955920,"results":"10","hashOfConfig":"11"},{"size":416,"mtime":1612863955920,"results":"12","hashOfConfig":"11"},{"size":534,"mtime":1612863955916,"results":"13","hashOfConfig":"11"},{"size":670,"mtime":1612863955916,"results":"14","hashOfConfig":"11"},{"size":269,"mtime":1613289864501,"results":"15","hashOfConfig":"11"},{"size":1013,"mtime":1612863955916,"results":"16","hashOfConfig":"11"},{"size":1617,"mtime":1612863955916,"results":"17","hashOfConfig":"11"},{"size":1172,"mtime":1612937374133,"results":"18","hashOfConfig":"11"},{"size":1816,"mtime":1613281359388,"results":"19","hashOfConfig":"11"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},"myvniu",{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/minh/Desktop/study-management/src/index.tsx",[],[],"/home/minh/Desktop/study-management/src/reportWebVitals.ts",[],"/home/minh/Desktop/study-management/src/App.tsx",[],"/home/minh/Desktop/study-management/src/components/Users.tsx",[],"/home/minh/Desktop/study-management/src/layouts/index.tsx",[],"/home/minh/Desktop/study-management/src/components/shared/button/PrimaryBtn.tsx",[],"/home/minh/Desktop/study-management/src/graphql/autogenerate/hooks.tsx",[],"/home/minh/Desktop/study-management/src/layouts/nav.tsx",[],"/home/minh/Desktop/study-management/src/components/adminNav.tsx",[]]
+[{"/home/minh/Desktop/study-management/src/index.tsx":"1","/home/minh/Desktop/study-management/src/reportWebVitals.ts":"2","/home/minh/Desktop/study-management/src/App.tsx":"3","/home/minh/Desktop/study-management/src/components/Users.tsx":"4","/home/minh/Desktop/study-management/src/layouts/index.tsx":"5","/home/minh/Desktop/study-management/src/components/shared/button/PrimaryBtn.tsx":"6","/home/minh/Desktop/study-management/src/graphql/autogenerate/hooks.tsx":"7","/home/minh/Desktop/study-management/src/layouts/nav.tsx":"8","/home/minh/Desktop/study-management/src/components/AdminNav.tsx":"9"},{"size":845,"mtime":1612863955920,"results":"10","hashOfConfig":"11"},{"size":416,"mtime":1612863955920,"results":"12","hashOfConfig":"11"},{"size":534,"mtime":1612863955916,"results":"13","hashOfConfig":"11"},{"size":670,"mtime":1612863955916,"results":"14","hashOfConfig":"11"},{"size":269,"mtime":1613291585025,"results":"15","hashOfConfig":"11"},{"size":1013,"mtime":1612863955916,"results":"16","hashOfConfig":"11"},{"size":1617,"mtime":1612863955916,"results":"17","hashOfConfig":"11"},{"size":1172,"mtime":1612937374133,"results":"18","hashOfConfig":"11"},{"size":1658,"mtime":1613291514128,"results":"19","hashOfConfig":"11"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},"myvniu",{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/minh/Desktop/study-management/src/index.tsx",[],[],"/home/minh/Desktop/study-management/src/reportWebVitals.ts",[],"/home/minh/Desktop/study-management/src/App.tsx",[],"/home/minh/Desktop/study-management/src/components/Users.tsx",[],"/home/minh/Desktop/study-management/src/layouts/index.tsx",[],"/home/minh/Desktop/study-management/src/components/shared/button/PrimaryBtn.tsx",[],"/home/minh/Desktop/study-management/src/graphql/autogenerate/hooks.tsx",[],"/home/minh/Desktop/study-management/src/layouts/nav.tsx",[],"/home/minh/Desktop/study-management/src/components/AdminNav.tsx",[]]
